Legacy Code/requirements.txt
Flask==2.3.2
googlemaps==4.10.0
python-dotenv==1.0.0
gunicorn==20.1.0